pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "signals"
version = "0.1.0"
description = "Markov chain-based market signal generator"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "jlg" }]
dependencies = [
    "yfinance>=0.2.40",
    "requests>=2.31",
    "pandas>=2.2",
    "numpy>=1.26",
    "pyarrow>=15.0",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "hmmlearn>=0.3",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7",
    "matplotlib>=3.8",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
signals = "signals.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["signals"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
